requirements.txt
aiogram==3.3.0
python-dotenv==1.0.1
openai==1.13.3
langchain==0.1.9
langchain-openai==0.0.5
pydantic==2.5.3
sqlalchemy==2.0.27
aiosqlite==0.19.0
chromadb==0.4.22
transformers==4.37.2
python-telegram-bot-pagination==0.0.2
tiktoken==0.5.2
faiss-cpu==1.7.4
unstructured==0.11.6
